# src/MCP.Views/mcp-chat-ui/openapi/backend.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MCP.WebApi | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5200/"
    }
  ],
  "paths": {
...
    "/mcp-servers/{serverName}": {
      "get": {
        "tags": [
          "McpServerConfiguration"
        ],
        "summary": "Get MCP server by name",
        "description": "Returns a single MCP server configuration item by name",
        "operationId": "GetMcpServerByName",
        "parameters": [
          {
            "name": "serverName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerConfigurationItem"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "McpServerConfiguration"
        ],
        "summary": "Create MCP server",
        "description": "Creates a new MCP server configuration item",
        "operationId": "CreateMcpServer",
        "parameters": [
          {
            "name": "serverName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerConfigurationItem"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerConfigurationItem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict"
          }
        }
      },
      "put": {
        "tags": [
          "McpServerConfiguration"
        ],
        "summary": "Update MCP server",
        "description": "Updates an existing MCP server configuration item",
        "operationId": "UpdateMcpServer",
        "parameters": [
          {
            "name": "serverName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerConfigurationItem"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerConfigurationItem"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "McpServerConfiguration"
        ],
        "summary": "Delete MCP server",
        "description": "Deletes an MCP server configuration item",
        "operationId": "DeleteMcpServer",
        "parameters": [
          {
            "name": "serverName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "McpServerConfigurationItem": {
        "required": [
          "category",
          "command",
          "args",
          "env",
          "url"
        ],
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "nullable": true
          },
          "command": {
            "type": "string",
            "nullable": true
          },
          "args": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "env": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "url": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/McpServerTransportType"
          }
        }
      },
      "McpServerTransportType": {
        "enum": [
          "stdio",
          "http"
        ],
        "default": "stdio"
      },
...
    }
  },
...
}